# static/newest-openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Generate Text API",
    "description": "Generate text, control stopping criteria like max_length/max_sentences",
    "version": "1"
  },
  "paths": {
...
    "/create-checkout-session": {
      "post": {
        "summary": "Create Checkout Session",
        "operationId": "create_checkout_session_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_checkout_session_create_checkout_session_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/create-user": {
      "post": {
        "summary": "Create User",
        "operationId": "create_user_api_create_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portal": {
      "get": {
        "summary": "Portal Redirect",
        "description": "redirect to the stripe customer portal",
        "operationId": "portal_redirect_portal_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Customer Id"
            },
            "name": "customer_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/get-user": {
      "post": {
        "summary": "Get User",
        "operationId": "get_user_api_get_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_checkout_session_create_checkout_session_post": {
        "title": "Body_create_checkout_session_create_checkout_session_post",
        "type": "object",
        "properties": {
          "uid": {
            "title": "Uid",
            "type": "string",
            "default": ""
          },
          "secret": {
            "title": "Secret",
            "type": "string",
            "default": ""
          }
        }
      },
      "CreateUserRequest": {
        "title": "CreateUserRequest",
        "required": [
          "email",
          "uid",
          "token"
        ],
        "type": "object",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "uid": {
            "title": "Uid",
            "type": "string"
          },
          "photoURL": {
            "title": "Photourl",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        }
      },
...
      "GetUserRequest": {
        "title": "GetUserRequest",
        "required": [
          "uid",
          "email"
        ],
        "type": "object",
        "properties": {
          "uid": {
            "title": "Uid",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}